true_mamba_experiments/requirements-true-mamba.txt
# Validated on this server with:
#   Python 3.10.20
#   torch 2.5.1+cu121 / torchvision 0.20.1+cu121 / torchaudio 2.5.1+cu121
# Install the torch stack, spconv-cu121, torch-scatter, causal-conv1d, and mamba-ssm
# with the dedicated commands documented in README.md and
# true_mamba_experiments/scripts/create_true_mamba_env.sh.

setuptools==75.8.0
numpy==1.23.5
hydra-core==1.3.2
omegaconf==2.3.0
pytorch-lightning==2.1.0
torchmetrics==0.11.0
pandas==2.1.1
scipy==1.11.3
matplotlib==3.10.8
pathlib2==2.3.7.post1
segmentation-models-pytorch==0.3.3
timm==0.9.2
efficientnet-pytorch==0.7.1
pretrainedmodels==0.7.4
wandb==0.13.5
ninja==1.13.0
